requirements.txt
# AI Rails TDD - Python Dependencies
# Core dependencies for Test-Driven Development workflow with n8n

# Web framework for webhook endpoints and human approval interface
fastapi==0.111.0
uvicorn[standard]==0.30.1

# HTTP client for API calls to Ollama and n8n
requests==2.32.3
httpx==0.27.0  # For async requests

# Ollama Python client for LLM interactions
ollama==0.3.3

# Data validation and serialization
pydantic==2.8.2
jsonschema==4.23.0

# Environment and configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Testing frameworks (for running generated tests)
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0

# Code quality tools (for validation)
black==24.4.2
ruff==0.5.5
mypy==1.11.1

# Utilities
rich==13.7.1  # For better console output
click==8.1.7  # For CLI tools

# Optional: For property-based testing
hypothesis==6.108.5

# Redis for state management
redis==5.0.1  # Async Redis client with connection pooling

# Vector search
upstash-vector==0.3.0  # Upstash Vector client for semantic search

# OpenAI for embeddings
openai==1.35.0  # OpenAI API client for generating embeddings